ai_agents/communication_agent/requirements.txt
# AISRi Communication Agent - Dependencies
# Python 3.11+ required

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Telegram Bot (will pull compatible httpx version)
python-telegram-bot==20.7

# HTTP Client - Let telegram-bot manage the version
# httpx is installed automatically by python-telegram-bot

# Supabase
supabase==2.3.4

# Scheduler
APScheduler==3.10.4

# Utilities
pydantic>=2.6.0
python-multipart==0.0.6
